Fixes: Openstack HA , mysql gets locked up during node failure test.
We have a cleanup script which cleans up CMON and Keystone tokens
every night. CMON produces a lot of logs (recorded into the database)
and hence the CMON DB in MySQL is pretty huge. When the CMON purge
script runs it results in data inconsistency between MySQL nodes
which is known to happen for DELETE operations in large MySQL tables.

The following change prevents this by doing the following -
* The time interval which CMON logs data is increased from 30 seconds
to 24 hours.
* Running CMON purge operation only in a node where CMON is running.
* Running cleanup every hour

Closes: 1414245

+++ contrail_provisioning/openstack/scripts/contrail-token-clean.sh
@@ -7,6 +7,8 @@
 source /etc/contrail/ha/cmon_param
 
 LOGFILE=/var/log/contrail/ha/token-cleanup.log
+MYIPS=$(ip addr show | sed -ne '/127.0.0.1/!{s/^[ \t]*inet[ \t]*\([0-9.]\+\)\/.*$/\1/p}')
+viponme=0
 mysql_user=keystone
 mysql_password=keystone
 mysql_host=$VIP
@@ -34,15 +36,23 @@ log_info_msg() {
     echo "$(timestamp): INFO: $msg" >> $LOGFILE
 }
 
-log_info_msg "keystone-cleaner::Starting token cleanup"
+for myip in $MYIPS
+ do
+  if [ $myip == $VIP ]; then
+     viponme=1
+     break
+  fi
+ done
+
+if [ $viponme -eq 1 ]; then
+    log_info_msg "keystone-cleaner::Starting token cleanup"
 
     mysql -u${mysql_user} -p${mysql_password} -h${mysql_host} -P${mysql_port} -e "USE keystone ; DELETE FROM token where expires <= convert_tz(now(),@@session.time_zone,'+01:00');"
     valid_token=$($mysql -u${mysql_user} -p${mysql_password} -h${mysql_host} -P${mysql_port} -e "USE keystone ; SELECT count(*) FROM token;")
     valid_token=$(echo $valid_token | awk '{print $2}')
 
-log_info_msg "keystone-cleaner::Finishing token cleanup, there are $valid_token valid tokens..."
+    log_info_msg "keystone-cleaner::Finishing token cleanup, there are $valid_token valid tokens..."
 
-    mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; ${cmon_stats_purge}"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table cmon_log;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table memory_usage_history;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table cpu_stats_history;"
@@ -51,6 +61,7 @@ log_info_msg "keystone-cleaner::Finishing token cleanup, there are $valid_token 
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table diskdata_history;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_global_statistics_history;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_statistics_history;"
+    mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_statistics_tm;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_advisor_history;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table expression_result_history;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_performance_results;"
@@ -65,8 +76,11 @@ log_info_msg "keystone-cleaner::Finishing token cleanup, there are $valid_token 
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table alarm_log;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_query_histogram;"
     mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; truncate table mysql_slow_queries;"
-
-log_info_msg "Purged cmon stats history"
+    mysql -u${cmon_user_pass} -p${cmon_user_pass} -h${mysql_host} -P${mysql_port} -e "use cmon; ${cmon_stats_purge}"
+    log_info_msg "Purged cmon stats history"
+else
+    log_info_msg "VIP is not in this node and hence not running the Keystone and CMON DB purge"
+fi
 
 find /var/log/contrail/ha/ -size +10240k -exec rm -f {} \;
 find /var/log/cmon.log -size +10240k -exec rm -f {} \;
